requirements.txt
Flask
torch
diffusers
pillow
werkzeug
transparent-background
numpy
